Added mongodb as adatabase
