+++ package.json
@@ -26,6 +26,7 @@
     "ag-grid-community": "^20.2.0",
     "ag-grid-enterprise": "^20.2.0",
     "core-js": "^2.5.4",
+    "mongodb": "^3.2.7",
     "rxjs": "~6.2.0",
     "zone.js": "~0.8.26"
   },
