--- projects/app2/src/styles.css
@@ -1,1 +1,0 @@
-/* You can add global styles to this file, and also import other style files */
