--- projects/utility/src/lib/utility.component.ts
@@ -1,19 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'lib-utility',
-  template: `
-    <p>
-      utility works!
-    </p>
-  `,
-  styles: []
-})
-export class UtilityComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
