--- projects/utility/src/lib/utility.module.ts
@@ -1,10 +1,0 @@
-import { NgModule } from '@angular/core';
-import { UtilityComponent } from './utility.component';
-
-@NgModule({
-  imports: [
-  ],
-  declarations: [UtilityComponent],
-  exports: [UtilityComponent]
-})
-export class UtilityModule { }
