--- projects/utility/src/lib/utility.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { UtilityService } from './utility.service';
-
-describe('UtilityService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: UtilityService = TestBed.get(UtilityService);
-    expect(service).toBeTruthy();
-  });
-});
