--- projects/utility/src/public_api.ts
@@ -1,7 +1,0 @@
-/*
- * Public API Surface of utility
- */
-
-export * from './lib/utility.service';
-export * from './lib/utility.component';
-export * from './lib/utility.module';
